moved validation functions to utils file
